Two preparatory steps that do not change behavior:

  Step 1 (128d2ff9): widen brk / mmap_next / ForkSnapshot from u32 to u64.
    The corresponding setter/getter signatures and exec_update_thread
    move to u64 too. Syscall ABIs (sys_sbrk i32->u32, sys_mmap u32->u32)
    are unchanged; the cast at the boundary is lossless as long as user
    space stays below 4 GiB.

  Step 2 (161ed11c): centralize the user-space VA layout in a new module
    kernel/src/memory/user_vmap.rs. Every constant describing program
    load address, DLIB region, heap limit, mmap region, user stack, and
    sigreturn trampoline now lives in one place. loader.rs, vma.rs,
    fork.rs, dll.rs, syscall/handlers/process.rs, and arch/x86/idt.rs
    were re-routed through it. Numerical values are unchanged.

The actual address move into the upper canonical half (à la Linux
x86_64) plus the relinking of every user binary is a separate
multi-day effort and is not part of this merge.

Boot verified in QEMU/KVM (virtio-vga) for both steps.

+++ kernel/src/task/scheduler/debug_trace.rs
@@ -811,8 +811,10 @@ pub fn thread_info_ex(target_tid: u32, buf_ptr: u64, size: u32) -> u32 {
     put_u32(&mut buf, 12, thread.cpu_ticks);
     put_u32(&mut buf, 16, thread.last_cpu as u32);
     put_u32(&mut buf, 20, thread.user_pages);
-    put_u32(&mut buf, 24, thread.brk);
-    put_u32(&mut buf, 28, thread.mmap_next);
+    // brk/mmap_next are u64 internally; layout still exposes the low 32 bits
+    // because user space currently lives below 4 GiB. Kept as u32 for ABI.
+    put_u32(&mut buf, 24, thread.brk as u32);
+    put_u32(&mut buf, 28, thread.mmap_next as u32);
     // Extract user-space RIP/RSP from kernel stack IRET frame (kernel_stack_top - 40/16)
     // context.rip/rsp contain kernel-internal addresses from context_switch.
     let ktop = thread.kernel_stack_top();
